scripts/requirements.txt
# Web Scraping
playwright==1.48.0
beautifulsoup4==4.12.3
requests==2.31.0

# Firebase
firebase-admin==6.5.0

# AI/ML
google-generativeai==0.8.3
python-dotenv==1.0.1

# Data Processing
pandas==2.2.3
